packages/mcp-bridge/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
mcp==1.9.4
pyyaml==6.0.2
